move options used to start HTTP server outside server package
